Replace old detector with new one
